refactor(api-markdown-documenter): Bundle `ApiItem`-related utilities into a namespace export (#17574)
Reduces clutter in the API surface by scoping utility functions.

New usage looks like the following:
```typescript
import { ApiItemUtilities } from "@fluid-tools/api-markdown-documenter";
...
const releaseTag = ApiItemUtilities.getReleaseTag(apiItem);
```

### Misc.

This PR also re-organizes utilities internally to differentiate
transformation-specific utilities from general-purpose utilities.

### Breaking Change

This PR removes a number of free functions from the global namespace.
Consumers will need to access via `ApiItemUtilities`.
