+++ tools/api-markdown-documenter/src/api-item-transforms/configuration/DocumentationSuiteOptions.ts
@@ -18,7 +18,7 @@ import {
 	getReleaseTag,
 	getUnscopedPackageName,
 	releaseTagToString,
-} from "../ApiItemUtilities";
+} from "../../utilities";
 
 /**
  * List of item kinds for which separate documents should be generated.
