+++ tools/api-markdown-documenter/src/api-item-transforms/helpers/TableHelpers.ts
@@ -28,16 +28,16 @@ import {
 	TableHeaderRowNode,
 	TableNode,
 } from "../../documentation-domain";
-import { injectSeparator } from "../../utilities";
 import {
 	ApiFunctionLike,
 	ApiModifier,
 	getDefaultValueBlock,
-	getLinkForApiItem,
 	getModifiers,
 	getReleaseTag,
+	injectSeparator,
 	isDeprecated,
-} from "../ApiItemUtilities";
+} from "../../utilities";
+import { getLinkForApiItem } from "../ApiItemTransformUtilities";
 import { transformTsdocSection } from "../TsdocNodeTransforms";
 import { getTsdocNodeTransformationOptions } from "../Utilities";
 import { ApiItemTransformationConfiguration } from "../configuration";
@@ -439,7 +439,7 @@ export function createPropertiesTable(
 		(apiItem) => getModifiers(apiItem, options?.modifiersToOmit).length > 0,
 	);
 	const hasDefaultValues = apiProperties.some(
-		(apiItem) => getDefaultValueBlock(apiItem, config) !== undefined,
+		(apiItem) => getDefaultValueBlock(apiItem, config.logger) !== undefined,
 	);
 
 	const headerRowCells: TableHeaderCellNode[] = [
@@ -619,7 +619,7 @@ export function createDefaultValueCell(
 ): TableBodyCellNode {
 	const tsdocNodeTransformOptions = getTsdocNodeTransformationOptions(apiItem, config);
 
-	const defaultValueSection = getDefaultValueBlock(apiItem, config);
+	const defaultValueSection = getDefaultValueBlock(apiItem, config.logger);
 
 	if (defaultValueSection === undefined) {
 		return TableBodyCellNode.Empty;
