+++ tools/api-markdown-documenter/src/utilities/index.ts
@@ -3,4 +3,7 @@
  * Licensed under the MIT License.
  */
 
+// All of the utilities here are meant to be used outside of this directory.
+// eslint-disable-next-line no-restricted-syntax
+export * from "./ApiItemUtilities";
 export { injectSeparator } from "./ArrayUtilities";
